fix: address PR review comments on DiagramCanvas
- fix measured preservation on resize: skip stale measured when consumer changed width/height to avoid edges anchored at wrong position
- split DiagramCanvas.js into Canvas.js (RF logic + shapes + propTypes) and DiagramCanvas.js (thin public wrapper), reducing DiagramCanvas.js from 325 to 38 lines
- export canvasPropTypes, DiagramNodeShape and DiagramEdgeShape from Canvas.js and reuse in DiagramCanvas.js
- rename translateMarker to formatMarker for clarity
- strip handleConfig from data before emitting onNodeClick (was leaking internal RF detail to consumers)
- add format.test.js with 26 tests covering mapNodesToRf, mapEdgesToRf, readNodeChanges and readEdgeChanges at 100% line coverage
